move dev module to separate file, update description
